update to 384.32948

+++ release/src/router/rc/rc_ipsec.c
@@ -388,7 +388,7 @@ void ipsec_prof_fill(int prof_idx, char *p_data, ipsec_prof_type_t prof_type)
 void ipsec_prof_fill_ext(int prof_idx, char *p_data, ipsec_prof_type_t prof_type)
 {
 	int i = 1;
-    char *p_end = NULL, *p_tmp = NULL, *ptr=NULL;;
+    char *p_end = NULL;
     p_end = p_data;
 
     /*encryption_p1_ext*/
@@ -516,8 +516,13 @@ void rc_ipsec_secrets_init()
 
 void rc_strongswan_conf_set()
 {
-    int rc = 0;
-    FILE *fp = NULL;
+	FILE *fp;
+	char *user;
+	int rc;
+
+	user = nvram_safe_get("http_username");
+	if (*user == '\0')
+		user = "admin";
 
     rc = pre_ipsec_samba_prof_set();
 //DBG(("ipsec_samba#\n"));
@@ -526,6 +531,7 @@ void rc_strongswan_conf_set()
                 "# Refer to the strongswan.conf(5) manpage for details\n#\n"
                 "# Configuration changes should be made in the included files"
                 "\ncharon {\n\n"
+                "  user = %s\n"
                 "  threads = %d\n"
                 "  send_vendor_id = yes\n"
                 "  duplicheck.enable = no\n"
@@ -536,7 +542,9 @@ void rc_strongswan_conf_set()
                 "  filelog {\n      /var/log/strongswan.charon.log {\n"
                 "        time_format = %%b %%e %%T\n        default = %d\n"
                 "        append = no\n        flush_line = yes\n"
-                "     }\n  }\n", nvram_get_int("ipsec_threads_num"), nvram_get_int("ipsec_log_level"));
+                "     }\n  }\n",
+                user,
+                nvram_get_int("ipsec_threads_num"), nvram_get_int("ipsec_log_level"));
     if(0 != rc){
         if(('n' != samba_prof.dns1[0]) && ('\0' != samba_prof.dns1[0])){
             fprintf(fp,"\n  dns1=%s\n", samba_prof.dns1);
@@ -1224,11 +1232,11 @@ void ipsec_conf_remote_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
     return;
 }
 
-char* get_ike_esp_bit_convert( char *str, int maxNum, int n, int type)
+char* get_ike_esp_bit_convert(char *str, int size, int maxNum, int n, int type)
 {
 	int i;
 	char tmpStr[12];
-	memset(str, 0, sizeof(str));
+	memset(str, 0, size);
 	for(i = 0; i < maxNum; i++){
 		if((n >> i) & 0x1 ){
 			if(type == FLAG_IKE_ENCRYPT)
@@ -1242,14 +1250,42 @@ char* get_ike_esp_bit_convert( char *str, int maxNum, int n, int type)
 	}
 	return str;
 }
+char* get_ike_esp_bit_convert1(char *str, int size, int n1, int n2, int n3)
+{
+	int i,j,k;
+	char tmpStr[SZ_MIN];
+	memset(str, 0, size);
+	memset(tmpStr, 0, sizeof(tmpStr));
+	for(i = 0; i < ENCRYPTION_TYPE_MAX_NUM; i++){
+		for(j = 0; j < HASH_TYPE_MAX_NUM; j++){
+			if(n3 !=0){
+				for(k = 0; k < DH_GROUP_MAX_NUM; k++){
+					if(((n1 >> i) & 0x1) && ((n2 >> j) & 0x1) && ((n3 >> k) & 0x1)){
+						sprintf(tmpStr, ",%s-%s-%s", encryp[i], hash[j], dh_group[k]);
+						strcat(str, tmpStr);
+					}
+				}
+			}
+			else{
+				if(((n1 >> i) & 0x1) && ((n2 >> j) & 0x1)){
+					sprintf(tmpStr, ",%s-%s", encryp[i], hash[j]);
+					strcat(str, tmpStr);
+				}
+			}
+		}
+	}
+	return str;
+}
 
 void ipsec_conf_phase1_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
 {
-	char str[128];
+	char str[1024];
+	memset(str, 0, sizeof(str));
     fprintf(fp, "  ikelifetime=%d\n", prof[prof_type][prof_idx].keylife_p1);
-	fprintf(fp, "  ike=%s", get_ike_esp_bit_convert(str, ENCRYPTION_TYPE_MAX_NUM, prof[prof_type][prof_idx].encryption_p1_ext, FLAG_IKE_ENCRYPT) + 1);
-	fprintf(fp, "%s", get_ike_esp_bit_convert(str, HASH_TYPE_MAX_NUM, prof[prof_type][prof_idx].hash_p1_ext, FLAG_ESP_HASH));
-	fprintf(fp, "%s\n", get_ike_esp_bit_convert(str, DH_GROUP_MAX_NUM, prof[prof_type][prof_idx].dh_group, FLAG_DH_GROUP));
+	fprintf(fp, "  ike=%s!\n", get_ike_esp_bit_convert1(str, sizeof(str), prof[prof_type][prof_idx].encryption_p1_ext, prof[prof_type][prof_idx].hash_p1_ext, prof[prof_type][prof_idx].dh_group) + 1);
+	//fprintf(fp, "  ike=%s", get_ike_esp_bit_convert(str, sizeof(str), ENCRYPTION_TYPE_MAX_NUM, prof[prof_type][prof_idx].encryption_p1_ext, FLAG_IKE_ENCRYPT) + 1);
+	//fprintf(fp, "%s", get_ike_esp_bit_convert(str, sizeof(str), HASH_TYPE_MAX_NUM, prof[prof_type][prof_idx].hash_p1_ext, FLAG_ESP_HASH));
+	//fprintf(fp, "%s\n", get_ike_esp_bit_convert(str, sizeof(str), DH_GROUP_MAX_NUM, prof[prof_type][prof_idx].dh_group, FLAG_DH_GROUP));
 
     /*if((ENCRYPTION_TYPE_MAX_NUM != prof[prof_type][prof_idx].encryption_p1) &&
        (HASH_TYPE_MAX_NUM != prof[prof_type][prof_idx].hash_p1)){
@@ -1275,8 +1311,9 @@ void ipsec_conf_phase2_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
 {
 	char str[128];
     fprintf(fp, "  keylife=%d\n", prof[prof_type][prof_idx].keylife_p2);
-	fprintf(fp, "  esp=%s", get_ike_esp_bit_convert(str, ENCRYPTION_TYPE_MAX_NUM, prof[prof_type][prof_idx].encryption_p2_ext, FLAG_IKE_ENCRYPT) + 1);
-	fprintf(fp, "%s\n", get_ike_esp_bit_convert(str, HASH_TYPE_MAX_NUM, prof[prof_type][prof_idx].hash_p2_ext, FLAG_ESP_HASH));
+	fprintf(fp, "  esp=%s!\n", get_ike_esp_bit_convert1(str, sizeof(str), prof[prof_type][prof_idx].encryption_p2_ext, prof[prof_type][prof_idx].hash_p2_ext, 0) + 1);
+	//fprintf(fp, "  esp=%s", get_ike_esp_bit_convert(str, sizeof(str), ENCRYPTION_TYPE_MAX_NUM, prof[prof_type][prof_idx].encryption_p2_ext, FLAG_IKE_ENCRYPT) + 1);
+	//fprintf(fp, "%s\n", get_ike_esp_bit_convert(str, sizeof(str), HASH_TYPE_MAX_NUM, prof[prof_type][prof_idx].hash_p2_ext, FLAG_ESP_HASH));
 
     /*if((ENCRYPTION_TYPE_MAX_NUM != prof[prof_type][prof_idx].encryption_p2) && 
        (HASH_TYPE_MAX_NUM != prof[prof_type][prof_idx].hash_p2)){
@@ -1386,11 +1423,11 @@ void rc_ipsec_topology_set()
     }
     return;
 }
-void rc_ipsec_nvram_convert_check()
+void rc_ipsec_nvram_convert_check(void)
 {
-	int i, rc = 0, prof_count = 0;
+	int i, prof_count = 0;
 	char buf[SZ_MIN], buf_ext[SZ_MIN];
-	char *nv=NULL, *nvp=NULL, *b=NULL;
+	char *nvp=NULL, *b=NULL;
 	for(prof_count = PROF_CLI; prof_count < PROF_ALL; prof_count++){
 	    for(i = 1; i <= MAX_PROF_NUM; i++){
 			if(PROF_SVR == prof_count){
@@ -1401,10 +1438,10 @@ void rc_ipsec_nvram_convert_check()
 				sprintf(&buf[0], "ipsec_profile_client_%d", i);
 				sprintf(&buf_ext[0], "ipsec_profile_client_%d_ext", i);
 			}
-			if(NULL != strcmp(nvram_safe_get(&buf[0]), "") && NULL == strcmp(nvram_safe_get(&buf_ext[0]),"")){
-				nv = nvp = strdup(nvram_safe_get(&buf[0]));
+			if(0 != strcmp(nvram_safe_get(&buf[0]), "") && 0 == strcmp(nvram_safe_get(&buf_ext[0]),"")){
+				nvp = strdup(nvram_safe_get(&buf[0]));
 				b = strsep(&nvp, ">");
-				if(NULL == strcmp(b, "4"))
+				if(0 == strcmp(b, "4"))
 					nvram_set(&buf_ext[0], "0>0>0>0>0");	/* none */
 				else
 					nvram_set(&buf_ext[0], "6>6>255>6>6");	/* 3des-aes128-sha1-sha256-modp768-modp1024-modp1536-modp2048-modp3072-modp4096-modp6144-modp8192 */
@@ -1413,30 +1450,32 @@ void rc_ipsec_nvram_convert_check()
 		}
 	}
 }
-void rc_ipsec_config_init()
+void rc_ipsec_config_init(void)
 {
     memset((ipsec_samba_t *)&samba_prof, 0, sizeof(ipsec_samba_t));
     memset((ipsec_prof_t *)&prof[0][0], 0, sizeof(ipsec_prof_t) * MAX_PROF_NUM);
-    memset((pki_ca_t *)&ca_tab[0], 0, sizeof(pki_ca_t) * CA_FILES_MAX_NUM);
+    //memset((pki_ca_t *)&ca_tab[0], 0, sizeof(pki_ca_t) * CA_FILES_MAX_NUM);
 	memset((ipsec_samba_t *)&pre_samba_prof, 0, sizeof(ipsec_samba_t));
-    system("cp -rf /usr/etc/* /tmp/etc/");
-    mkdir("/jffs/ca_files", 0777);
+	if(!d_exists("/etc/ipsec.d") || !d_exists("/etc/strongswan.d"))
+		system("cp -rf /usr/etc/* /tmp/etc/");
+    //mkdir("/jffs/ca_files", 0777);
     /*ipsec.conf init*/    
     rc_ipsec_conf_default_init();
     rc_ipsec_psk_xauth_rw_init();
     /*ipsec.secrets init*/
-    rc_ipsec_set(IPSEC_INIT,PROF_ALL);
+    if(nvram_get_int("ipsec_server_enable") || nvram_get_int("ipsec_client_enable"))
+		rc_ipsec_set(IPSEC_INIT,PROF_ALL);
     //rc_ipsec_secrets_set();
     //rc_ipsec_conf_set();
     /*ipsec pki shell script default generate*/
     //rc_ipsec_ca_default_gen();
     //rc_ipsec_pki_gen_exec();
-    rc_ipsec_ca_init();
+    //rc_ipsec_ca_init();
     /*ca import*/
     //rc_ipsec_ca_import();
     return;
 }
-void rc_set_ipsec_stack_block_size()
+void rc_set_ipsec_stack_block_size(void)
 {
 	char command[64];
 	if(NULL != nvram_safe_get("ipsec_stack_block_size")){
@@ -1471,7 +1510,7 @@ void get_bitmap_scan(int *cur_bitmap_en)
 	}
 	return;
 }
-void run_ipsec_firewall_scripts()
+void run_ipsec_firewall_scripts(void)
 {
 	char *argv[3];
 	argv[0] = "/bin/sh";
@@ -1734,7 +1773,7 @@ void rc_ipsec_topology_set_XML()
 					strcat_r(flag_buf," require-cfgmode allow-cfgmode",flag_buf);
 				if(IKE_AGGRESSIVE_MODE == prof[prof_count][i].exchange)
 					strcat_r(flag_buf," aggressive-mode",flag_buf);
-				if(NULL == strcmp(prof[prof_count][i].tun_type,"transport"))
+				if(0 == strcmp(prof[prof_count][i].tun_type,"transport"))
 					strcat_r(flag_buf," transport",flag_buf);
 				
 				if(0 != strlen(flag_buf))
@@ -1792,7 +1831,7 @@ void rc_ipsec_topology_set_XML()
 				
 				if(0 != prof[prof_count][i].keylife_p2)
 					fprintf(fp,"      <life type=\"seconds\">%d</life>\n", prof[prof_count][i].keylife_p2);	
-				if(NULL != strcmp(prof[prof_count][i].remote_gateway, "null"))
+				if(0 != strcmp(prof[prof_count][i].remote_gateway, "null"))
 					fprintf(fp,"        <peer>%s</peer>\n", prof[prof_count][i].remote_gateway);
 				fprintf(fp,"      <local-ip>%s</local-ip>\n", prof[prof_count][i].local_pub_ip);
 				fprintf(fp,"      <ike-versions>%d</ike-versions>\n", prof[prof_count][i].ike);
